Now drive content page uses layout and handles the case when a user checks a folder to which they do not have acccess
