+++ src/app/f/[folderId]/layout.tsx
@@ -1,0 +1,75 @@
+"use client";
+
+import { useEffect, useState } from "react";
+import { Button } from "~/components/ui/button";
+import { Input } from "~/components/ui/input";
+import { Search, Moon, Sun } from "lucide-react";
+import { SignedIn, SignedOut, SignInButton, UserButton } from "@clerk/nextjs";
+
+export default function RootLayout({
+  children,
+}: Readonly<{ children: React.ReactNode }>) {
+  const [searchQuery, setSearchQuery] = useState("");
+  const [isDark, setIsDark] = useState(true);
+
+  useEffect(() => {
+    document.documentElement.classList.add("dark");
+  }, []);
+
+  const toggleTheme = () => {
+    const newIsDark = !isDark;
+    setIsDark(newIsDark);
+
+    if (newIsDark) {
+      document.documentElement.classList.remove("light");
+      document.documentElement.classList.add("dark");
+    } else {
+      document.documentElement.classList.remove("dark");
+      document.documentElement.classList.add("light");
+    }
+  };
+
+  return (
+    <div className="bg-background flex min-h-screen flex-col">
+      <header className="border-border bg-card/50 border-b px-8 py-6 shadow-sm backdrop-blur-sm">
+        <div className="flex items-center justify-between">
+          <div className="flex items-center space-x-8">
+            <h1 className="text-primary text-2xl font-bold tracking-tight">
+              Drive Clone
+            </h1>
+          </div>
+          <div className="flex items-center space-x-6">
+            <Search className="text-muted-foreground absolute ms-4 h-4 w-4 -translate-y-1/2 transform" />
+            <Input
+              placeholder="Search in Drive"
+              value={searchQuery}
+              onChange={(e) => setSearchQuery(e.target.value)}
+              className="border-border bg-background focus:border-primary h-11 w-96 pl-12 transition-colors duration-200"
+            />
+            <Button
+              variant="outline"
+              size="icon"
+              onClick={toggleTheme}
+              className="border-border hover:bg-muted h-11 w-11 bg-transparent transition-colors duration-200"
+            >
+              {isDark ? (
+                <Sun className="h-4 w-4" />
+              ) : (
+                <Moon className="h-4 w-4" />
+              )}
+            </Button>
+            <SignedOut>
+              <SignInButton />
+            </SignedOut>
+            <SignedIn>
+              <UserButton />
+            </SignedIn>
+          </div>
+        </div>
+      </header>
+      <main className="mx- px- flex flex-grow flex-col px-32 py-8">
+        {children}
+      </main>
+    </div>
+  );
+}
